Remove obsolete code and variables submodule
