+++ fair/register.py
@@ -221,106 +221,3 @@ def fetch_registrations(
         _stored_objects.append(_file_url)
     
     return _stored_objects
-            
-def subst_registrations(local_uri: str, input_config: typing.Dict):
-    """As part of 'run' substitute listings in config for working configuration
-
-    Parameters
-    ----------
-    uri: str
-        endpoint of the local registry
-    input_config : typing.Dict
-        input user configuration
-
-    Returns
-    -------
-    typing.Dict
-        new dictionary with 'register' removed and replaced with local objects
-    """
-    _new_cfg = input_config.copy()
-
-    # If there is no register key then abort substitutions
-    if 'register' not in _new_cfg:
-        return _new_cfg
-
-    # Remove register as it will be replaced
-    del _new_cfg['register']
-
-    if 'read' not in _new_cfg:
-        _new_cfg['read'] = []
-
-    _registrations = input_config['register']
-
-    for reg in _registrations:
-        if all(t not in reg for t in ['data_product', 'external_object']):
-            raise fdp_exc.UserConfigError(
-                "Expected either 'data_product' or 'external_object' in 'register' item"
-            )
-
-        if 'version' not in reg:
-            raise fdp_exc.InternalError(
-
-            )
-
-        # If an external object fetch the relevant data_product first
-        if 'external_object' in reg:
-            # TODO Not enough info to make unique
-            _ext_data = {
-                "version": reg['version'],
-                "title": reg['title']
-            }
-            _results = fdp_req.get(
-                local_uri,
-                'external_object',
-                params=_ext_data
-            )
-
-            if len(_results) > 1 or not _results:
-                raise fdp_exc.InternalError(
-                    f"Expected one external_object for '{_ext_data}', "
-                    f"got {len(_results)}"
-                )
-
-            if 'data_product' not in _results[0]:
-                raise fdp_exc.RegistryError(
-                    f"Expected external_object '{reg['external_object']}' "
-                    "to have a data_product"
-                )
-            _data_product_url = _results[0]['data_product']
-            _results[0] = fdp_req.url_get(_data_product_url)
-        else: # data product
-            # TODO Not enough info to make unique
-            _data = {
-                "version": reg['version'],
-                "name": reg['data_product']
-            }
-
-            if 'namespace_name' in reg:
-                _data['namespace'] = reg['namespace_name'] 
-
-            _results = fdp_req.get(
-                local_uri,
-                'data_product',
-                params=_data
-            )
-
-            if len(_results) > 1 or not _results:
-                raise fdp_exc.InternalError(
-                    f"Expected one result for {_data}, got {len(_results)}"
-                )
-
-        _namespace_url = _results[0]['namespace']
-        _namespace_data = fdp_req.url_get(_namespace_url)
-
-        _object_data = {
-            'data_product': _results[0]['name'],
-            'use': {
-                'version': _results[0]['version'],
-                'namespace': _namespace_data['name'],
-            }
-        }
-
-    _new_cfg['read'].append(_object_data)
-
-    return _new_cfg
-
