+++ pyproject.toml
@@ -74,6 +74,7 @@ pytest-cov = "^2.12.1"
 requests-mock = "^1.9.3"
 deepdiff = "^5.5.0"
 pytest-virtualenv = "^1.7.0"
+loremipsum = "^1.0.5"
 
 [tool.poetry.scripts]
 fair = 'fair.cli:cli'
