+++ pytest.ini
@@ -12,3 +12,4 @@ markers=
     user_config: tests for the 'user_config' submodule
     staging: tests for the 'staging' submodule
     cli: tests for the CLI itself
+    variables: tests for 'parsing.variables' submodule
